pyproject.toml
[tool.poetry]
name = "course-7"
version = "0.1.0"
description = ""
authors = ["Ksenia-Kostyuk <hatan2235@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
